src/test/javascript/spec/app/entities/sub-sis-grupo/sub-sis-grupo.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { SubSisGrupoService } from 'app/entities/sub-sis-grupo/sub-sis-grupo.service';
import { ISubSisGrupo, SubSisGrupo } from 'app/shared/model/sub-sis-grupo.model';

describe('Service Tests', () => {
  describe('SubSisGrupo Service', () => {
    let injector: TestBed;
    let service: SubSisGrupoService;
    let httpMock: HttpTestingController;
    let elemDefault: ISubSisGrupo;
    let expectedResult: ISubSisGrupo | ISubSisGrupo[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(SubSisGrupoService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new SubSisGrupo(0, 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a SubSisGrupo', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new SubSisGrupo()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a SubSisGrupo', () => {
        const returnedFromService = Object.assign(
          {
            gidDesigna: 'BBBBBB',
            gidGrupo: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of SubSisGrupo', () => {
        const returnedFromService = Object.assign(
          {
            gidDesigna: 'BBBBBB',
            gidGrupo: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a SubSisGrupo', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
